Fix basic issues with the execution flow in the fibers

+++ engine/console/simpleLexer.h
@@ -761,7 +761,7 @@ private:
       {
          // integer
          Token t = make(TokenType::INTCONST);
-         t.ivalue = std::stoi(s);
+         t.ivalue = std::stoll(s);
          return t;
       }
    }
